relocated viewport initialization
